# api/openapi/restfleet-v1.yaml
openapi: 3.1.0
info:
  title: RestFleet API
  version: 0.1.0
  description: RestFleet V1 contract.
servers:
  - url: /
security:
  - sessionCookie: []
paths:
...
  /api/v1/bootstrap/status:
    get:
      operationId: bootstrapStatus
      security: []
      responses:
        "200":
          description: Whether the first administrator must be created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapStatus"
        "503":
          $ref: "#/components/responses/Problem"
  /api/v1/bootstrap:
    post:
      operationId: bootstrap
      security: []
      parameters:
        - name: X-RestFleet-Bootstrap-Token
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 4096
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BootstrapRequest"
      responses:
        "201":
          description: First administrator and authenticated session created.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /api/v1/auth/login:
    post:
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated session created.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /api/v1/auth/logout:
    post:
      operationId: logout
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204":
          description: Session revoked and cookies cleared.
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /api/v1/auth/session:
    get:
      operationId: session
      responses:
        "200":
          description: Current authenticated session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /api/v1/dashboard/summary:
    get:
      operationId: dashboardSummary
      responses:
        "200":
          description: Empty-capable control-plane summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
...
  /api/v1/hosts:
    get:
      operationId: listHosts
      responses:
        "200":
          description: Hosts visible to the current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostList"
        "401":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
    post:
      operationId: createHost
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostCreate"
      responses:
        "201":
          description: Host created in PENDING state.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Host"
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/v1/hosts/{host_id}:
    parameters:
      - $ref: "#/components/parameters/HostId"
    get:
      operationId: getHost
      responses:
        "200":
          description: Host detail.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Host"
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
    patch:
      operationId: updateHost
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostPatch"
      responses:
        "200":
          description: Host updated.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Host"
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/v1/hosts/{host_id}/disable:
    parameters:
      - $ref: "#/components/parameters/HostId"
    post:
      operationId: disableHost
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Host disabled.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Host"
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/v1/hosts/{host_id}/enable:
    parameters:
      - $ref: "#/components/parameters/HostId"
    post:
      operationId: enableHost
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Host enabled.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Host"
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/v1/hosts/{host_id}/agents:
    parameters:
      - $ref: "#/components/parameters/HostId"
    get:
      operationId: listHostAgents
      responses:
        "200":
          description: Agent identities associated with the Host.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentList"
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/v1/hosts/{host_id}/inventory:
    parameters:
      - $ref: "#/components/parameters/HostId"
    get:
      operationId: getHostInventory
      responses:
        "200":
          description: Latest inventory reported by an Agent associated with the Host.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentInventory"
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/v1/hosts/{host_id}/enrollment-tokens:
    parameters:
      - $ref: "#/components/parameters/HostId"
    get:
      operationId: listEnrollmentTokens
      responses:
        "200":
          description: Enrollment token metadata. Bearer tokens are never returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentTokenList"
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
    post:
      operationId: createEnrollmentToken
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentTokenCreate"
      responses:
        "201":
          description: Enrollment token created. The bearer value appears only in this response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentTokenCreated"
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/v1/enrollment-tokens/{token_id}:
    parameters:
      - $ref: "#/components/parameters/TokenId"
    delete:
      operationId: revokeEnrollmentToken
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204":
          description: Unused enrollment token revoked.
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/v1/agent-enrollment:
    post:
      operationId: enrollAgent
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentEnrollmentRequest"
      responses:
        "201":
          description: Agent identity bundle. The private key is never returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentEnrollmentResponse"
        "400": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/v1/agents/{agent_id}:
    parameters:
      - $ref: "#/components/parameters/AgentId"
    get:
      operationId: getAgent
      responses:
        "200":
          description: Agent identity and connection metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/v1/agents/{agent_id}/revoke:
    parameters:
      - $ref: "#/components/parameters/AgentId"
    post:
      operationId: revokeAgent
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRevoke"
      responses:
        "200":
          description: Agent and all currently valid certificates revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
...
  /api/v1/storage-credentials/{credential_id}:
    parameters:
      - name: credential_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getStorageCredential
      responses:
        "200":
          description: Credential metadata; ETag contains the resource revision.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StorageCredential" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: restfleet_session
  parameters:
    HostId:
      name: host_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: '^"?[1-9][0-9]*"?$'
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 4096
  responses:
    Problem:
      description: RFC 9457 problem response.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    BootstrapStatus:
      type: object
      additionalProperties: false
      required: [bootstrap_required]
      properties:
        bootstrap_required:
          type: boolean
    BootstrapRequest:
      type: object
      additionalProperties: false
      required: [username, display_name, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
          pattern: "^[A-Za-z0-9._-]+$"
        display_name:
          type: string
          minLength: 1
          maxLength: 128
        password:
          type: string
          minLength: 12
          maxLength: 1024
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64
        password:
          type: string
          minLength: 1
          maxLength: 1024
    Session:
      type: object
      additionalProperties: false
      required: [user, idle_expires_at, absolute_expires_at]
      properties:
        user:
          $ref: "#/components/schemas/User"
        idle_expires_at:
          type: string
          format: date-time
        absolute_expires_at:
          type: string
          format: date-time
...
    HostList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Host" }
...